jflte: loki: support flashing loki'd recovery images

+++ releasetools/loki-recovery.sh
@@ -4,43 +4,45 @@
 # See here for more information on loki: https://github.com/djrbliss/loki
 #
 
-export C=/data/local/tmp/loki_tmpdir
 RECSIZE=$1
 RECSHA1=$2
 BOOTSIZE=$3
 BOOTSHA1=$4
 
-mkdir -p $C
-
 egrep -q -f /system/etc/loki_bootloaders /proc/cmdline
 if [ $? -eq 0 ]; then
   need_lok=1
+  export C=/data/local/tmp/loki_tmpdir
+  rm -rf $C
+  mkdir -p $C
+  dd if=/dev/block/platform/msm_sdcc.1/by-name/recovery of=$C/recovery.lok
+  /system/bin/loki_tool unlok $C/recovery.lok $C/recovery
+else
+  export C=/dev/block/platform/msm_sdcc.1/by-name/
 fi
 
-if ! applypatch -c EMMC:$C/recovery.img:$RECSIZE:$RECSHA1; then
-  log -t recovery "Installing new recovery image"
-
+if ! applypatch -c EMMC:$C/recovery:$RECSIZE:$RECSHA1; then
   if [ $need_lok -eq 1 ]; then
-    dd if=/dev/block/platform/msm_sdcc.1/by-name/aboot of=$C/aboot.img
-    dd if=/dev/block/platform/msm_sdcc.1/by-name/recovery of=$C/recovery.lok
+    log -t recovery "recovery is outdated. unloki-ing all the things"
     dd if=/dev/block/platform/msm_sdcc.1/by-name/boot of=$C/boot.lok
-    /system/bin/loki_tool unlok $C/recovery.lok $C/recovery.img
-    /system/bin/loki_tool unlok $C/boot.lok $C/boot.img
-  else
-    dd if=/dev/block/platform/msm_sdcc.1/by-name/boot of=$C/boot.img
+    dd if=/dev/block/platform/msm_sdcc.1/by-name/aboot of=$C/aboot.img
+    /system/bin/loki_tool unlok $C/boot.lok $C/boot
   fi
 
-  applypatch -b /system/etc/recovery-resource.dat EMMC:$C/boot.img:$BOOTSIZE:$BOOTSHA1 EMMC:$C/recovery.img $RECSHA1 $RECSIZE $BOOTSHA1:/system/recovery-from-boot.p || exit 1
+  log -t recovery "Installing new recovery image"
+  applypatch -b /system/etc/recovery-resource.dat EMMC:$C/boot:$BOOTSIZE:$BOOTSHA1 EMMC:$C/recovery $RECSHA1 $RECSIZE $BOOTSHA1:/system/recovery-from-boot.p || exit 1
 
   if [ $need_lok -eq 1 ]; then
-    /system/bin/loki_tool patch recovery $C/aboot.img $C/recovery.img $C/recovery.lok || exit 1
+    /system/bin/loki_tool patch recovery $C/aboot.img $C/recovery $C/recovery.lok || exit 1
     /system/bin/loki_tool flash recovery $C/recovery.lok || exit 1
-  else
-    dd if=/tmp/boot.img of=/dev/block/platform/msm_sdcc.1/by-name/recovery
   fi
+
 else
   log -t recovery "Recovery image already installed"
 fi
 
-rm -rf $C
+if [ $need_lok -eq 1 ]; then
+  rm -rf $C
+fi
+
 exit 0
